requirements.txt

Django==4.2.16
typing_extensions==4.12.2
djangorestframework~=3.15.2
requests~=2.32.3
pydub~=0.25.1
SpeechRecognition~=3.11.0
pygame~=2.6.1
